requirements.txt
psycopg2-binary>=2.9.0
pandas>=2.0.0
geopandas>=0.13.0
shapely>=2.0.0
sqlalchemy>=2.0.0
numpy>=1.24.0
netCDF4>=1.6.0
xarray>=2023.0.0
matplotlib>=3.7.0
requests>=2.31.0
lxml>=4.9.0